+++ src/system/run.js
@@ -126,9 +126,8 @@ var Run = {
         wait: system.wait_scale,
       });
 
-      var verbose_command = options.command_verbose === true;
       var docker_opt = system.daemonOptions(options, image.Config);
-      var container  = yield lazy.docker.run(system.image.name, docker_opt.command, docker_opt, verbose_command);
+      var container  = yield lazy.docker.run(system.image.name, docker_opt.command, docker_opt);
 
       if (options.wait) {
         var first_tcp = _.find((docker_opt.ports_orderly || []), (data) => {